requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fonttools==4.40.0
idna==3.4
imutils==0.5.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.125
urllib3==2.0.3
